+++ compat/external_consumer/lib/alloy_external_consumer.dart
@@ -9,6 +9,8 @@ export 'package:alloy_external_consumer/src/device_info.dart';
 export 'package:alloy_external_consumer/src/diagnostics.dart';
 export 'package:alloy_external_consumer/src/platform_module.dart';
 export 'package:alloy_external_consumer/src/report.dart';
+export 'package:alloy_external_consumer/src/reporter.dart';
 export 'package:alloy_external_consumer/src/repository.dart';
 export 'package:alloy_external_consumer/src/search_index.dart';
 export 'package:alloy_external_consumer/src/system_clock.dart';
+export 'package:alloy_external_consumer/src/telemetry.dart';